Fixing #1764 - a property's 'sync' method is never considered a no-op.
*This is a behaviour change.*

If the property does not return an event name, then one is generated
based on the property name.

Previously, the 'sync' method could return nil and it would be considered
a noop, but if you need a noop, then you need to modify your 'insync?' method
to return 'true' in the noop cases.

Also modifying all of the builtin types that didn't handle this explicitly or
returned nil in 'sync'.  There should be no behaviour change in any of them.

Signed-off-by: Luke Kanies <luke@madstop.com>